+++ src/api/core/operators/audio_vec_embedding.py
