+++ src/api/core/operators/test_audio_tag.py
@@ -1,12 +1,12 @@
 import unittest
 from unittest.case import skip
-import audio_tag
+import audio_vec_embedding
 
 class Test(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         # initialize operator
-        audio_tag.initialize(param={})
+        audio_vec_embedding.initialize(param={})
     
     @classmethod
     def tearDownClass(cls):
@@ -15,5 +15,5 @@ class Test(unittest.TestCase):
 
     def test_sample_audio_from_disk(self):
         audio_file_path = r'sample_data/audio.wav'
-        audio_emb = audio_tag.run(audio_file_path)
+        audio_emb = audio_vec_embedding.run(audio_file_path)
         self.assertEqual(2048, len(audio_emb))
